+++ src/dipdup/config/starknet_subsquid.py
@@ -5,18 +5,18 @@ from typing import Literal
 from pydantic import ConfigDict
 from pydantic.dataclasses import dataclass
 
-from dipdup.config import DatasourceConfig
 from dipdup.config import HttpConfig
 from dipdup.config import Url
+from dipdup.config._subsquid import SubsquidDatasourceConfig
 
 
 @dataclass(config=ConfigDict(extra='forbid', defer_build=True), kw_only=True)
-class StarknetSubsquidDatasourceConfig(DatasourceConfig):
+class StarknetSubsquidDatasourceConfig(SubsquidDatasourceConfig):
     """Subsquid datasource config
 
     :param kind: always 'starknet.subsquid'
     :param url: URL of Subsquid Network API
-    :param api_key: API key for self-hosted Subsquid gateways (required by `v2.archive.subsquid.io` since 2026-05-19)
+    :param api_key: API key
     :param http: HTTP client configuration
     """
 
@@ -24,12 +24,3 @@ class StarknetSubsquidDatasourceConfig(DatasourceConfig):
     url: Url
     api_key: str | None = None
     http: HttpConfig | None = None
-
-    @property
-    def merge_subscriptions(self) -> bool:
-        return False
-
-    @property
-    def rollback_depth(self) -> int:
-        # NOTE: Subsquid data is always finalized
-        return 0